Support for one handed operation
Users with normal sized thumbs (that means me!) find it difficult to type
words using letters in the top right corner (words with "P" and "O"
letters). This patch tries to make these words easier to type.

At the moment this patch is ambidextrous and should be helpful for
lefties out there (but i don't have any way to test)

(and there are probably some magic improvements stuffed here and there)