Variable weight between curve & prediction engines
We interpolate the coefficient based on speed and scores

If you type very fast prediction engine will have more weight and is
more likely to correct typing mistakes.
On the contrary, if you want to type a rare or difficult to type word,
type it a bit slowly.
It seems to cope well with real world situation, but this will require a lot of
testing :-)

Coefficients and low/high thresholds are auto-tuned to adapt to user