Prediction scores now have a "score type"
This allow to know when two scores are comparable (they share the same
type). Otherwise the one with the greater type always wins.

For exemple a score achieved with word trigrams ("S3") are always more
significant than those coming from cluster trigrams ("C3") or word bigrams("S2")

This is used in the backtracking function to know if a word replacement
is valid.