Multi-touch swiping initiated
Not sure about this, but this is a tentative attempt for a multi-touch
keyboard (maybe like Keymonk or Nintype).

It may prove useful for the new landscape split keyboard or for the
much awaited tablet.

At the moment it's completely untested (it's just the minimal code
change to validate that i've not broken non regression tests and that
there is no performance impact)

It's in a separate branch and not scheduled for upcoming release.