Remove limit on cluster n-grams count
Clustering was implemented to solve the n-gram sparsity issue in order
to get good negative filtering (i.e. find words that are *not* likely to be
found in a given context). Pruning makes this property weaker, so this
commit just removes this limit.

Cluster depth has already been chosen to fit our size requirements so
there won't be any significant language files size increase (and it will
stay far below the new limit).